requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pymysql==1.1.1
simple-openai==0.9.0
numpy==2.1.3
scikit-learn==1.5.2
tenacity==9.0.0
PyYAML==6.0.2
